Safari-611.1.10/LayoutTests/accessibility/aria-multiselectable-grid-expected.txt

This tests that aria-multiselectable is exposed correctly for grids.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS gridMultiselectableUnspecified.isMultiSelectable is true
PASS gridMultiselectableTrue.isMultiSelectable is true
PASS gridMultiselectableFalse.isMultiSelectable is false
PASS successfullyParsed is true

TEST COMPLETE
